// components/Admin/Products/Forms/types.tsx
export interface Product {
  product_id: number;
  product_name: string;
  product_sku: string;
  product_description: string;
  long_description: string;
  product_price: number;
  product_quantity: number;
  product_discount: number;
  product_status: "draft" | "pending" | "approved";
  tags: string[];
  main_image: string;
  thumbnails: {
    thumbnail1: string;
    thumbnail2: string;
    thumbnail3: string;
    thumbnail4: string;
    thumbnail5: string;
  }[];
  category_id: string;
  brand_id: string;
  brand_name: string;
  brand_image: string;
  specifications: {
    specification_name: string;
    specification_value: string;
    category_id: string;
  }[];
  suppliers: {
    supplier_id?: number;
    supplier_name?: string;
    supplier_email?: string;
    supplier_phone_number?: string;
    supplier_location?: string;
    isNew?: boolean;
  }[];
}

export interface ProductFormProps {
  initialData?: Product;
}

export interface Supplier {
  supplier_id?: number;
  supplier_name?: string;
  supplier_email?: string;
  supplier_phone_number?: string;
  supplier_location?: string;
  isNew?: boolean;
}

export interface Brand {
  brand_id: number;
  brand_name: string;
  brand_image: File | null | string;
}

export interface Category {
  category_id: string;
  category_name: string;
}